+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## TBD
+
+* Avoid polling when detecting ANRs by invoking JNI from SIGQUIT handler
+  [#741](https://github.com/bugsnag/bugsnag-android/pull/741)
+
 ## 4.22.2 (2020-01-06)
 
 ### Bug fixes